- 부가서비스 카테고리 추가

+++ daema-api/src/main/java/com/daema/rest/commgmt/dto/OrganizationMemberDto.java
@@ -30,5 +30,5 @@ public class OrganizationMemberDto {
 	private long storeId;
 	private long orgId;
 	private String userStatus;
-	private Integer[] roleIds;
+	private Long[] roleIds;
 }
